+++ include/minishell.h
@@ -76,7 +76,6 @@ void	free_data(t_data *data);
 //pipex funciones:
 int		ft_pipex(t_data data);
 int		ft_exec_cmd(t_data *data, t_cmd *node, int cmd_number);
-//utils_pipex funciones:
 int 	ft_cmdlist_size(t_cmd *cmd_list);
 int		ft_is_builtin(t_data *data, char *str);
 int		ft_built_in(t_data *data, t_cmd *node);
