+++ src/pipex/list_utils.c
@@ -1,0 +1,2 @@
+#include "../../include/minishell.h"
+
